[Refactor]: Improve folder structure
